requirements.txt
youtube-transcript-api==0.4.1
python-docx==0.8.11
pytube==12.0.0
requests==2.26.0


